Smart_Plant_Platform/backend/disease/temp/requirements.txt
onnxruntime>=1.15.0
numpy>=1.21.0
Pillow>=9.0.0 